+++ tools/cliutils/build.gradle
@@ -11,6 +11,7 @@ dependencies {
     compile "info.picocli:picocli:$picocli_version"
     compile "com.jcabi:jcabi-manifests:$jcabi_manifests_version"
     compile "org.slf4j:slf4j-api:$slf4j_version"
+    compile "org.apache.logging.log4j:log4j-slf4j-impl:$log4j_version"
 
     // JAnsi: for drawing things to the terminal in nicely coloured ways.
     compile "org.fusesource.jansi:jansi:$jansi_version"
